+++ packages/protocol/src/index.ts
@@ -3,3 +3,9 @@ export { DEPLOYMENTS } from "./addresses";
 export type { Deployment } from "./addresses";
 export type { Offer, CommitInfo, PendingCommit, CommitResult } from "./types";
 export { loadDeployment } from "./deployment-loader";
+
+export { ENTRY_POINT_V06, computeUserOpHash } from "./userop";
+export type { UserOperation } from "./userop";
+
+export { readEscrowConstants, readRegistryConstants } from "./constants";
+export type { EscrowConstants, RegistryConstants } from "./constants";